+++ src/common_traits/iterator/mod.rs
@@ -5,10 +5,10 @@ mod into_iter;
 mod iter;
 mod iter_mut;
 mod iter_mut_rev;
+mod iter_of_slices;
 mod iter_ptr;
 mod iter_ptr_bwd;
 mod iter_rev;
-mod iter_slices;
 mod iter_slices_of_con;
 mod reductions;
 
@@ -19,8 +19,8 @@ pub use into_iter::IntoIter;
 pub use iter::Iter;
 pub use iter_mut::IterMut;
 pub use iter_mut_rev::IterMutRev;
+pub use iter_of_slices::{IterOfSlices, SliceBorrowAsMut, SliceBorrowAsRef};
 pub use iter_ptr::IterPtr;
 pub use iter_ptr_bwd::IterPtrBackward;
 pub use iter_rev::IterRev;
-pub use iter_slices::{IterOfSlices, SliceBorrowAsMut, SliceBorrowAsRef, SliceBorrowKind};
 pub use iter_slices_of_con::IterSlicesOfCon;
